Adding https2http frontend
Adding https2http frontend at port 81. The purpose of this frontend is to handle traffic which was original https,
but forwarded to http (for example, when SSL is terminated at AWS ELB).

This frontend never redirects to https (since it is assumed that traffic came from https already) and sets
Strict-Transport-Security header if configured to do so.

Also, making http frontend to redirect to https  if "ingress.kubernetes.io/force-ssl-redirect" annotation is set to "true".