Fix health check of dynamic reload
Dynamic reload updates IP and port to the new endpoint without updating the health check port number. This could be fixed either updating the health check port or removing the port from the server configuration in the template. Since we don’t have an option to use a different port to the health check, the port was removed from the server configuration. HAProxy uses the same port of the service if not declared.