Override stop signal from upstream image
The upstream image is used to start HAProxy itself as pid1. Its stop signal was changed to SIGUSR1 which does a soft reload. Our pid1 is the controller which only listen SIGTERM (docker stop) and SIGINT (^C).

Changing the stop signal in the container image as a way to document what the controller process is expecting.