Fix overlapping configs on shared frontend
An overlap happens whenever a `/path` of a domain has a configuration, eg whitelist or body-size limit, and a `/path/sub` doesn’t have such configuration. Most of match is made with path_beg which will also match the subpath uri. This change adds a `!{ subpath }` acl to fix this behavior.