Skip check if the instance need to be reloaded
A `haproxy -c` is made in order to certify that the new configuration file is valid. Such check is already performed when the instance is reloaded, without stopping the old instance if the check fails, so skipping check in order to save some time and proc on big config files.