Change annotation storage from struct to map
The very first reading and storage of ingress and service annotation values was on pseudo-structured types. This has the benefit of saving some typing but has some issues, mainly:

1) Cannot distinguish between a defined annotation with default value and an undefined annotation - one cannot eg overwrite a default value with zero (int), false (bool) or empty (string)
2) Cannot void collisions of the same annotation with distinct values to distinct paths of the same backend

This change move from structs to maps, allowing to overwrite default values with empty one. It is however missing changes in haproxy types and template in order to fix backend annotations collisions.