Add annotation map to haproxy object converter
Some configurations has two or more annotations which compose a single object. Using HSTS as an example, there are four annotations that might be declared on configmap, one or more ingress objects or service object whose values could diverge on a subpath.

This mapper implementation builds an abstraction that simplifies to group annotations in order to populate internal haproxy objects properly.