Validate and normalize annotation values
A validation will be made in the annotation data input, which will avoid an invalid data from being ever stored in the map, and if not stored, the default value will be used instead.

Validation will also normalize the value, so eg an int value "0" or "00" will always be used as "0", will use a single map entry in the backend configuration, which means less options to HAProxy worry about.