Add var-namespace annotation
A HAProxy var txn.namespace can be configured with the k8s namespace owner of the service which is the target of the request. This can be used on http logs. var-namespace is a boolean annotation that defaults to false, which means txn.namespace shouldn't be created.

This change only defines the annotation. The haproxy entity was already implemented.