Fix service port lookup
Use the backend port to find the correct ServicePort object from the service. The ServicePort name is used to match the Endpoint port name. Finally, the endpoint port number is the right one. ServicePort name is optional only if there is only one port, so if ServicePort.Name == "", the Endpoint port match, otherwise the svcPort and epPort names should match.