Fix port number lookup of terminating pods
Terminating pods are used on DrainSupport mode. They don't have endpoints so need their own func to find the correct container port if a named port is being used.

The service's target port should be a valid port number, which means it is the container port number, or should be a valid named port declared in the container. If the port is nether a valid number nor a declared named port, the endpoint of that pod will not be added.