Make sni optional if a certificate is optional and is not provided
The sni extension is mandatory if a client certificate authentication is used. However, if the certificate is optional, the request was falling back to the default backend if sni isn't provided.

This commit changes how requests to domains protected by authentication via client cert works: if a client certificate is optional and the request does not provide one, the host header will be used to match a backend. If a certificate is mandatory or a certificate is sent, the sni extension continue to be mandatory and will be used instead.

This should be merged to v0.8.