Use 421 response if sni and headers does not match
Browsers reuse a stablished TLS connection if a wildcard or multi DNS certificate matches more than one domain. This is a problem if the new request uses client cert authentication, whose validation happens on the TLS handshake.

A 421 HTTP response makes the browser create a new TLS connection. This response will be used if the sni extension (from the TLS handshake) and the host header (from the request) does not match.

* https://www.mail-archive.com/haproxy@formilux.org/msg29789.html
* https://www.mail-archive.com/haproxy@formilux.org/msg33799.html