Add leader election
A leader elector is used to select which ingress-controller instance should perform operations that should be done just once - by the leader. A configmap object in the k8s cluster is used to control which instance is the leader.

Use cases:

* status update: the status of the ingress object should be updated by only one instance
* acme signer: sign of new certificates should be done by only one instance

The old controller has its own leader elector to update ingress status. That leader elector will be deprecated before v0.9-beta.