Improve fronting proxy config
Fixes https redirect scheme and port sharing with http-port, which defaults to 80. Name changed from `https-to-http-port` to `fronting-proxy-port` which better represents the functionality. The old name is now an alias to the new one.

The implementation was moved from https frontend to http due to some advantages:

* Easier to share the same http port without side effect
* There is less `if`s in the http frontend if compared to the https one

Missing var-namespace and max-body-size implementations which wasn't in the http frontend.

Should be merged to v0.8.