Fix host match with a port number
Request processing are made using the `base` sample (host+path) and sometimes only the Host header is used. Some http clients add the port number in the header, giving something like `domain.local:80` where only the `domain.local` part should be used to match ingress hostnames. This conversion is done using regsub() that matches and removes `:[0-9]+/`. The problem here is the trailing slash: it only exists in the `base` sample and doesn't exist if the host header is used.

Should be merged to v0.8