Add initial-weight config key
The `initial-weight` configuration key defines the static weight configuration of each backend server, which will define the initial weight configuration. The initial weight configuration is used on agent-check: an agent can change the weight of a server defining a percentage value of the initial weight.

Without this option, servers always start with weight 1 (one) and empty slots start with weight 0 (zero). If trying to update the weight via agent, in the former, the weight cannot be downgraded without removing the server from the balance, and the upgrade is always at least duplicating the weight. In the latter the weight cannot be changed. Iow weight via agent cannot be used.

Should be back ported to v0.8.