fix hostname match of domains with client cert auth
concat() converter was declared wrong in two ways: using the first argument which is a constant instead of a dynamic arg, and using a fetch sample instead of a var name. This commit fixes both.

There is also an improvement on sni parsing: a regex was being used to remove a possible port number after the domain name. This port number is a possibility in the Host header but it is not in the sni extension, so the regex was removed.