fix txn.namespace on http requests
Requests on http frontend wasn't filling txn.namespace var. This haproxy var can be used in http logs to reference the k8s namespace owner of the ingress object which declares the domain and path of the request. The namespace mapping should also be changed to an internal global scope in order to have visibility from the http frontend.

This should be merged to v0.8.