do ssl-redirect only if tls declares the hostname
Ingress.spec.tls declares hostnames/domains and an optional secret with crt+key used to ssl-offload https requests on that domains. If a secret name is not declared or is invalid, the default crt+key is used instead.

This commit fixes a misbehaviour when ssl-redirect is true, declared or inherited from configmap, without a TLS being declared. Now ssl-redirect only return 302 redirect if ingress.spec.tls declares the hostname/domain.

Should be merged to v0.8.