use the found match pattern
The string `_nomatch` was used everywhere as a fallback if some pattern isn't found in a map. This used to work nice provided that the source sample could be found.

There is however a special use case when the sni extension isn't provided and the domain has optional client cert: the response is 404 due to the fetch sample failed. This doesn't affect v0.8 because the failed sample is new in v0.9 - see concat().

This change removes the default value on map*() converters, so it can fail, won't create the target variable and `-m found` can be used instead.