initial implementation of acme-v2 protocol
This initial version has the following features:

- a client that authorises new certificate sign
- a server that answers challenges to the requested domains
- a work queue in order to serialize requests, avoiding flood the acme-v2 server
- a leader election controls which haproxy-ingress instance is the leader and should sign certificates
- a rate limiting queue which exponentially increase the time between requests of failed domains

Now there are two event loops in place: one which parses ingress resources, build haproxy config and, dynamic update or reload the instance. Now another event loop is also running to process acme challenges and certificate generation/storage - if acme is enabled. There is also a goroutine which does a full check looking for invalid or expiring certificates - by default every 24h.

There are some pieces of the code that need to be configured before starting the ingress event loop, because of that some configurations need to be done via command-line config, which means read only. Most of them are related with event loop control.

The haproxy config "incorrectly" stores the acme data - now the config struct has a new session "external state", which means configurations that shouldn't trigger a haproxy reload.

The leader elector defines which haproxy-ingress instance should start the acme-v2 challenge and sign new certificates. Every time the leader changes, the new leader does a full check in order to see if any certificate generation is still pending - this can happens on restarts or when the actual leader dies/is stopped.