add remove func to the work queue
An item will continue in the work queue until its processing returns nil as the error object. Remove() is a way to stop waiting it to successfully execute, removing it from the queue.

This is used on invalid domains on the acme queue - if the dev remove the domain from the ingress, the old request will be removed from the queue. A new request would be added if the domain is just changed.