remove certs removed from the config
An invalid cert request will continue in the error queue until it is successfully processed, which means forever for an invalid request. This change removes a request in the error queue whenever the request is removed from the configuration. A changed request will be removed (old version) and added again (new version).