remove unused acme tokens
Acme challenges happen configuring a token which should be served when the acme environment calls a URL in the domain being validated. Every ingress instance should know this token, so because of that it is stored in k8s.

Old tokens wasn't being removed after being used, this change fixes this behaviour.