add cors preflight as Lua service
A cors preflight is entirely managed by haproxy, no need to connect a backend server. A Lua service was already being used to skip the request to the backend. This change is moving all the headers and status code from haproxy cfg to the service, on behalf of a cleaner config.