remove v0.7 controller
All the code within the `common` package is deprecated. This is an inheritance from the old times of the so-called "ingress controllers" where haproxy-ingress was just a flavor. v0.8 reimplemented most of the reused functionality and now the old codebase is being removed. There was some missing and not removed pieces which will be reimplemented in a future version: listers/cache, ssl stores and status update.