add frontend to the internal prometheus exporter
Configures a frontend and a listening port (defaults to 9100) to scrape metrics from an internal Prometheus exporter implementation.

Added two new global config keys to configure binding ip and port number. Port as zero removes the frontend.

The prometheus exporter is slower and more verbose than the csv exporter, added a note about this in the doc.