add metrics to haproxy response time
Add the set server command to the haproxy_response_time histogram. Set server is used by dynamic updates.

The approach used is to receive a callback in HAProxyCommand(), so the func itself will register the elapsed time and register the metric. The caller, in this case instance and dynupdate, need to know an instance of Metric intf, whose `HAProxy<xxx>ResponseTime()` func should be used as a callback implementation.