postpone rotateConfig on haproxyUpdate
Rotate config, currently, takes the config object on curConfig and copy it to oldConfig, and also assign nil to the curConfig reference. The oldConfig reference has two meanings:

1) it is used to compare with a new curConfig in order to perform dynamic updates
2) if oldConfig != nil it means the controller has already started haproxy

The 2) is used by `CalcIdleMetric()`, which cannot connect to the admin socket on its early executions on big configurations. Such configurations take some seconds to start haproxy for the first time.