Change unix sockets user to haproxy
The internal ACME server (answer ACME challenges) and the http frontend (only if internal fronting TCP proxy is needed) uses unix sockets to the haproxy process communicate with. If the controller starts as root but `use-haproxy-user` is declared, haproxy wouldn't have permission to write to the socket. Now all sockets has `haproxy` as owner and `0600` as its permission, even if the process continue to run as root.