Resync cert_expire metric and crt list
Every domain+cn combination generates a new time series. A time series will continue to exist and to be scraped until it is removed via prometheus client or the controller is restarted. When domain or cn is changed, a new time series is created without removing the old one. This update compares old and new configuration in order to remove time series from domain+cn combinations that doesn't exist anymore.