Sort tcp services by name and port
HAProxy Ingress compares old and cur configuration objects to detect if an apiserver notification is in fact a configuration change. This will only work properly if slices are always in the same order. TCP services were being sorted only by name, which will lead to random restarts if only the port number changes. This happens because the source of the data is a `configmap.data[]` which is a map.

Fixes #493