create frontends even without ingress
Fixes a wrong assumption that a frontend need at least one ingress to be configured. Instead, a frontend should always be configured and issue a 404 page, since any host could be matched.

This also fixes configurations whose ingress objects have empty host - empty or wildcard `*` hosts are used to configure or overwrite the default backend without create a proper host entry, and were falling back in the same problem of not create the frontend.