Remove list of frontends
A frontend+bind list were historically used to separate distinct client timeouts, distinct `ca-file` if auth-tls is used, and also distinct hostnames in pre-historical versions. Client timeout was just moved to global scope, which allows to use a single frontend and a single bind config despite the ingress and controller configuration.

This update doesn't have any functional change - it is merely removing the legacy code used to build a list of frontends with a list of binds. Such simplifications will speedup the development of performance tunings.

A L6 frontend is still needed and only configured if at least one hostname uses ssl-passthrough.

TCP services wasn't changed and will continue to be configured as usual.