move frontend maps to the frontend entity
Frontend group was converted to a link of maps after the last changes. This update renames the func and the old frontend group entity, and also moves the maps link to the frontend entity. The new `Write(Frontend|Backend)Maps()` funcs doesn't change the model state and should continue like this.