move listers and informers to the new controller
This reimplementation is functionally identical compared with the old one, but in the new controller package and in a decoupled way. The new work queue implementation is also being used to enqueue changes notified by the informers.

Current status updater and TLS certs are tight coupled with the old controller and with the old listers and informers. Their tests was designed with that coupling in mind and they would need to be rewritten after moving the lister. Because of that, status updater and TLS certs tests were removed. They need to be migrated instead of maintained.

The old store was just some facades on top of the listers, they were merged with the (new) controller cache.

All other changes are just removing legacy types and fields, and updating old code to call/reference the new controller/cache/lister.