fix cleaning cache of changed objects
A list of changed objects is updated by Notify(), which listen events from informers. On every ingress parsing, this list is processed and cleared, waiting to be populated again on new events. The list of changed services wasn't being cleared, which means that every time any service was created or updated, the same hosts and backends tracked by that service was being synchronised and rewritten to disk every time any watched k8s object changes. Such "leak" would invalidate all the gains of partial parsing and backend-shards on a long running controller.