add match type support to hostsmap
MatchType allows to configure how a hostname and/or path received will match against a map/list of hostnames and/or paths used to route requests or apply constraints. This is a pre requisite to path as regex and PathType from the k8s 1.18 ingress spec.

HostsMap was rewritten from scratch and its rewrite changed the naming pattern (again) of the HAProxy maps. Most of the changes in the template and instance_test refers to this naming change.