remove empty maps
Up to v0.10 map files was just simple and regex, where simple was always created. This is being changed now that we can have up to four files (and counting) per filter type. So lets render, write and load map files only when needed, adjusting the configuration file to behave just like it was empty - without the need to create and load it.

At this moment only the current implementation was migrated: Begin and Regex match.