add Exact and Prefix match types
A new match type in haproxy is done using a new string match or a new converter derivative. The model configuration is abstract, only need to create a new `MatchType` constant and use it as a path attribute.

Most of the changes are adding new `http-request set-var()` keywords reading a new `map()` converter derivative - `map()` when exact match and `map_dir()` when prefix match. Begin was already implemented and using `map_beg()`, regex was already implemented as well (wildcards and alias-regex annotation) and using `map_reg()`.

Most of the test changes reflect the impact of adding a new acl on `map_beg()` due to the following chain:

- `setvar(v) fetch,map()`
- `setvar(v) fetch,map_dir() if !{ var(v) -m found }`
- `setvar(v) fetch,map_beg() if !{ var(v) -m found }`
- `setvar(v) fetch,map_reg() if !{ var(v) -m found }`

The `map_beg()` line didn't need a `-m found` acl before `map()` and `map_dir()` implementation, because of that his commit does not have new tests - they will be added in a new commit on behalf of visibility of the changes.