add PathLink type
A PathLink uniquely identifies a hostname and a path. It is used during the configuration building to identify the target(s) of a configuration, which the config file generator uses to build keywords and ACLs.

The old implementation just concatenate hostname and path, which doesn't work well on regex match, so this refactor is a pre requisite to properly implement it.