remove implicit boundaries from regex match
Change the regex match path type to not use implicit begin-of-line and end-of-line commands, resulting in a more intuitive configuration and compatible with other web servers. The problem of this approach is that hostname and path are concatenated and used together to match domain and path in a single step. The expected behavior was preserved changing the presence or absence of the begin/end of line between hostname and path to a proper regex.