add case sensitive match to exact, prefix and regex
The only match type up to now was begin, which is case insensitive. This was implemented lowering all `host`, `path` and `base` fetch methods as well as the hostnames and paths declared in the ingress object - `map()` converter does not have a case insensitive version.

Now lower is only used in the hostname and map files used to match the `begin` path type, other path types are case sensitive and are compared as is.