fix tracking and partial parsing of spec.backend
Most of the tracking stuff was missing on Ingress.Spec.Backend, resulting in an inconsistent state and sometimes a broken config. Hosts.defaultHost pointer was also removed, now the default `*` host is part of the Hosts.items map and inherits all the Hosts' partial parsing implementation, eg RemoveAll, ItemsAdd and Shrink.