Add 'session-cookie-preserve' backend annotation
The 'session-cookie-preserve` annotation is able to be used when cookie
affinity is enabled. The purpose is to be able to intelligently control
dynamic updating in such a way where it avoids causing an update that
would make the cookie value for a server no longer match an Enabled
server (mismatch on disabling a server is fine).

The reason cookie values matching their assigned backend is important,
is because the "preserve" mode of HAProxy allows servers to emit a
"Set-Cookie" header to effectively assign their own server. Because of
this, the specific cookie value assigned to each backend server must be
consistent and known by both HAProxy as well as the backend server
itself.

This change makes usage of "preserve" mode in "session-cookie-keywords"
no longer (or at least, less) desirable, but it should still be allowed
for backward compatibility and maximum customizability, therefore a
warning will simply be logged when it is used in that way.