Add 'session-cookie-value-strategy' annotation
This new backend annotation allows the way that a cookie value for a
backend server is calculated to be controlled more precisely.

Prior to this, the value of a cookie for a backend server was always set
to the Name of an endpoint, which was in turn controlled by
'backend-server-naming'. While that gives a few options for changing the
cookie value, the downside was that it required changing the name of the
server. Additionally, there are possible cookie values which would make
poor backend server names (such as a guid/uuid) but may make good cookie
values.

The new annotation defaults to setting the naming strategy to the Name
of the server, which is the existing behaviour. In addition, it adds the
ability to set the value to a Pod UID.