remove instance cmd options
Instance cmd, currently haproxy and reload, was configured in the
most external controller code and now it's not compatible with the
new external haproxy option and remote commands. This will need
another refactor in the future, moving all magic strings to a single
place.

A new field `fake` is now in place. Unit tests defines this field as
true in order to the instance code know that there is no haproxy to
be reloaded. Btw this will also need a refactor in the future, using
mocks instead of ifs.