add compatibility with master-worker and external haproxy
Add new global configuration keys that can be used to adjust the
config file to an external haproxy deployment:

* `username` and `groupname` that configures a non root user that
should own the running haproxy process;
* `external-has-lua` that currently loads `auth-request.lua`, this
script would crash haproxy if loaded and O.S. doesn't have Lua
socket installed;
* `master-exit-on-failure` that configs such option in the master
worker mode;
* `use-haproxy-user` continue to behave just like v0.9-v0.11, but
now it's an alias to configure username and groupname as haproxy.

hatypes.External.IsExternal value is not being updated yet, its
value depends on the master socket being configured - this is a task
of a future commit.

haproxy's `unix-bind` keyword now only changes the user and group of
unix sockets if a non root user is used, there is no reason to
change the socket user if the haproxy process owner isn't changed -
and it also simplifies the embedded/external configuration
compatibility.

oauth2 was also updated to not be configured if `external-has-lua`
wasn't enabled, otherwise it would fail haproxy bootstrap.