improve old and new backend comparison
Old and new backends are compared whenever a direct or indirect related
ingress, service or endpoint changes. This comparison improves CPU usage
avoiding cfg bulding if configurations match, and haproxy reload if a
dynamic update can be made.

A static comparison check if all but endpoints match. This comparison
need to ignore internal controller attributes related with endpoint,
otherwise any change eg to the endpoints order would count as a backend
change, which is not right.

`pathConfig` was added in Per Path Backend Config and since then it was
incorrectly firing some backend cfg parsings. This commit remove it from
the comparison.