fix dynamic update of the default backend
Since v0.11, all cluster updates try to be partially applied in the
in-memory model. This is addressed building an object tracking, followed
by purging the old state and recreating these same objects in the new
state. In short ingress objects track hostnames, services and secrets,
and services track endpoints.

Default backends wasn't being updated because it exposes a service
without using an ingress object, so endpoints are deployed on a full
synchronization, including de first one, and wasn't tracker anymore.
This update creates a dummy tracking point to bind the default backend
endpoints and the haproxy model.

This update also needed to manage an issue with the backend ID: the
default backend is renamed to `_default_backend` when assigned as such,
and changed back when unassiged. This unassignment happens when the
backend is purged in order to have its internal state updated, and such
asynchrony should be fixed in the dynupdate. Now dynupdate compares
backends based on an internal and immutable state.