configure default backend to not change backend ID
The default backend configured via a command-line option has a
special name since ages. This comes from a time when it was in fact
a special entity created apart of the regular services referenced by
ingress objects. Actually the default backend is also a regular
backend that needs to be tracked and updated, and such special name
is incompatible with future refactors planned in the backend entity.
Iow the special name does not worth the effort anymore.