fix basic auth backend tracking
Basic auth users and passwords are stored in secret resources. We're
properly tracking secret->userlist in the event of a secret update, and
backends are also tracked but only when a new userlist is created.

A proper partial parsing need to track either a host or a backend from
the haproxy model. If an untracked backend uses a tracked userlist,
changes to this userlist will remove it without recreate, which will
lead to failures in the configuration parsing.

This update enforces that a backend is always tracked in the event of
a secret update, ensuring that partial parsing works properly.