read ingress class to validate an ingress resource
According to the [official doc](https://kubernetes.io/docs/concepts/services-networking/ingress/#ingress-class),
an ingress is valid for an instance of a controller if the deprecated
annotation `kubernetes.io/ingress.class` references the controller or the
ingress resource references an ingress class that references it. In the
case of a conflict, annotation wins.

This implementation still consider `ignore-ingress-without-class`, but
it's default value should be changed to `true` shortly in order to follow
what the documentation states.