remove key parsing of acme secret
The acme secret is used 1) to save a valid x509 certificate signed by a
certificate authority using acme protocol and its private key, and 2) by
haproxy to perform a TLS handshake. Both crt and key are used by haproxy
but the acme signer doesn't need to read the key - only the crt need to
be read in order to know if the certificate need to be reissued.

The current key parsing implementation only recognizes RSA keys. It
doesn't worth the effort to implement parsing of other key types because
the key isn't used at all by the acme signer.