add ingress class reading and tracking
Add read of ingress class name and tracking the class with the hostname.
Ingress resources was already filtered by the cache, so we don't need to
validate it. Missing class names are tracked to rebuild the hostname when
the class is added in the future.

The class is only tracked, it isn't stored anywhere because `parameters`
isn't implemented yet.