wait ingress class initialization
Ingress informers now reads ingress class to decide if an ingress
resource should be processed in that controller instance. Because of
that, ingress class lister need to be properly initialized to avoid
random and incorrect "ingress class not found" due to race during
controller initialization.

This change adds another WaitForCacheSync() to wait ingress class
initializes and before start ingress initialization.