add ingress class related command-line option
A new command-line option `--controller-class` allows to change the
controller name that HAProxy Ingress should look for in the
IngressClass resources. If declared, the provided string is added as a
sub path of the controller name.

This commit also changes the behavior of tracking orphan ingress
resources - ingresses that declares neither the class name nor the class
annotation. Up to now HAProxy Ingress was adding such orphans in the
configuration, now the default behavior is to ignore them, which
strictly follows the spec. A new command-line option
`--watch-ingress-without-class` changes this behavior and also
deprecates `--ignore-ingress-without-class`.

refs:

* https://v1-18.docs.kubernetes.io/docs/concepts/services-networking/ingress/#ingress-class
* https://v1-18.docs.kubernetes.io/docs/reference/generated/kubernetes-api/v1.18/#ingressspec-v1beta1-networking-k8s-io
* https://v1-18.docs.kubernetes.io/docs/reference/generated/kubernetes-api/v1.18/#ingressclassspec-v1beta1-networking-k8s-io