use default cert only if sni isn't declared
Default or fallback certificate is defined in haproxy declaring it as
the first certificate. It should be used in TLS handshakes when SNI is
not provided or if the provided one does not have a match.

HAProxy Ingress uses `crt-list` to better implements Ingress spec: a
valid certificate is not one that its CN or SAN has the provided SNI,
but instead that one the user declares in the spec.tls hostname.

There is however a scenario which trigger the wrong path: if the default
certificate is valid for a provided SNI, haproxy will use it in the
handshake instead of a declared certificate - this happens because the
default certificate, being the first one, is validated first.

This commit adds a `!*` in the snifilter of the default certificate.
This will instruct haproxy to not add any of its domains - CN and SAN -
in the SNI, so it won't match any domain, except if a valid one cannot
be found. This is exactly what a default certificate is all about.