fix duplication of userlist
A basic auth's userlist can be reused if the same password secret is
reused in more than one Ingress, or even on the same Ingress with two or
more paths. The userlist was being duplicated due to a missing find func.
This doesn't break the configuration because all the userlists have the
same content and haproxy complains but doesn't refuse to start.