independently configure rules and TLS
Since the very first controller version, the TLS spec was only configured
if a corresponding hostname was found in the rules of the same ingress
resource. This commit changes this behavior and reads the TLS array as an
independent configuration.

Such update changes backward compatibility since a misconfigured TLS
would be silently ignored. However the new behavior better implements
the ingress spec.