only notifies ConfigMap updates if data changes
ConfigMap now can be used as the IngressClass' Parameters target.
Because of that ConfigMaps in the controller namespace are also being
watched. However other ConfigMaps that changes periodically was starting
the ingress parsing - eg the cm used for leader election which is update
every few seconds.

This update checks if a ConfigMap's data attribute was changed,
otherwise a reconciliation isn't started.