clear the crt expire gauge when full sync
A full syncing happens whenever a global configuration changes, the
default crt is updated or, if for any reason the lister event sends
invalid data or it cannot be successfully processed. Whenever this
happens, domains that was removed in the mean time continue to populate
the crt expire gauge.

This commit rebuilds the gauge if the haproxy config need to be rebuilt.
Note that, although such clean ups shouldn't be very often, when they
happen the gauge will have incomplete data during a very small amount
of time.