fix first conciliation if external haproxy is not running
The very first config conciliation might happen before haproxy is
properly running and listening to its master socket. This can happen if
eg the haproxy image should be download and the controller image is
present. The first conciliation will now wait until the admin socket can
successfully answer `show info`, and will only exit to the loop after
send a command without error or if the controller receives a sigterm.