remove match type empty
Match type `empty` was a work around to reuse some functions from an
old HostsMap implementation. After the refactor to the path type
support, only `crt-list` was using it and without a valuable reuse.
This small refactor is now a pre-requisite to fix the path type
priority. Crt list need its own template file, but for now we'll
touch as few code as possible because we need to merge this to v0.12
and v0.11.