add sni and verifyhost to secure connections
Add options to securely connect over untrusted networks and avoid
man in the middle attack. Without these options only date and the
certificate authority that issued the server certificate can be
verified. SNI static or dynamically configures the SNI TLS extension to
the backend server, and also verifies the name of the server certificate
if a ca bundle is provided. Verifyhost does the same without changing
the SNI extension and only accepts a static hostname.