add dynamic update of frontend's tls certificate
Evolve dynupdate to recognize changes to the frontend's TLS certificate
and apply it using the runtime api. The update also happens even if was
detected that a reload should also be performed, so the configuration
will be as updated as possible if for any reason the reload fails.

This code updates the minimal HAProxy version to 2.2. HAProxy 2.1 also
implements TLS cert update, but this version cannot update crt-lists
with filters - which we use in the default certificate.